Move the Decimal struct into its own file
If, in the future, additional structs need to be added, they can all be nicely
separated into their own files.  Also keeps the entry point clean.
